LayoutTests/imported/w3c/web-platform-tests/html/dom/elements/global-attributes/dir-auto-dynamic-changes.window-expected.txt


PASS dynamic insertion of RTL text in a child element
PASS dir=auto changes for content insertion and removal, in and out of document
PASS dir=auto changes for slot reassignment
FAIL text changes affecting both slot and ancestor with dir=auto assert_false: slot after first text change expected false got true
PASS dynamic changes to subtrees excluded as a result of the dir attribute
PASS dynamic changes inside of non-HTML elements
FAIL slotted non-HTML elements assert_true: state after first change (slot) expected true got false
PASS slotted non-HTML elements after dynamically assigning dir=auto, and dir attribute ignored on non-HTML elements
PASS dir=auto ancestor considers text in subtree after removing dir=ltr from it
PASS Slotted content affects multiple dir=auto slots
PASS Removing slotted content resets direction on dir=auto slot
FAIL Removing child of slotted content changes direction on dir=auto slot assert_equals: slot is reset to ltr expected "ltr" but got "rtl"
PASS Child directionality gets updated when dir=auto is set on parent
FAIL dir=auto slot is updated by text in value of input element children assert_equals: expected "ltr" but got "rtl"
FAIL dir=auto slot is updated if input stops being auto-directionality form-associated assert_equals: expected "ltr" but got "rtl"
PASS slot provides updated directionality from host to a dir=auto container
א
א א
א

 abc
